+++ src/components/CodeEditor.tsx
@@ -1,13 +1,13 @@
-import React, { useState } from 'react';
-import MonacoEditor from 'react-monaco-editor';
+import React, { useRef, useState } from 'react';
+import MonacoEditor from "@monaco-editor/react";
 
 
 const CodeEditor: React.FC = () => {  
-  const [code, setCode] = useState<string>('console.log("Hello, TypeScript!");');
+  const [code, setCode] = React.useState("//Just a comment");
 
-  const handleCodeChange = (newValue: string) => {
-    setCode(newValue);
-  };
+  function handleEditorChange(value: any, event: any) {
+    setCode(value);
+  }
 
   const executeCode = () => {
     const iframe = document.createElement('iframe');
@@ -23,30 +23,36 @@ const CodeEditor: React.FC = () => {
     document.body.removeChild(iframe);
   };
 
+  function handleEditorDidMount(editor: any, monaco: any) {
+    console.log('onMount: the editor instance:', editor);
+    console.log('onMount: the monaco instance:', monaco);
+  }
+
+  function handleEditorWillMount(monaco: any) {
+    console.log('beforeMount: the monaco instance:', monaco);
+  }
+
+  function handleEditorValidation(markers: any) {
+    // model markers
+    // markers.forEach(marker => console.log('onValidate:', marker.message));
+  }
+
   return (
     <div>
       <div style={{ marginBottom: '20px' }}>
         <button onClick={executeCode}>Run</button>
       </div>
       <MonacoEditor
-        width="800"
-        height="600"
-        language="typescript"
-        theme="vs-dark"
-        value={code}
-        options={
-          {
-            selectOnLineNumbers: true,
-            roundedSelection: false,
-            readOnly: false,
-            cursorStyle: "line",
-            automaticLayout: false
-          }
-        }
-        onChange={handleCodeChange}
+        height="90vh"
+        defaultLanguage="typescript"
+        defaultValue="// some comment"
+        onChange={handleEditorChange}
+        onMount={handleEditorDidMount}
+        beforeMount={handleEditorWillMount}
+        onValidate={handleEditorValidation}
       />
     </div>
-  );
+  )
 };
 
 export default CodeEditor;
